Add namelist option config_halo_exch_method in &development for selecting halo method
The new namelist option can be set to either 'mpas_dmpar' to select the use of
existing group halo exchange routines from the mpas_dmpar module, or it can be
set to 'mpas_halo' to select the use of routines from the mpas_halo module.

At present, the actual calls to halo exchange routines are not affected by
the setting of config_halo_exch_method: the mpas_dmpar halo exchange routines
are always called.