Correct halo exchanges
This commit adds working halo exchange code. The issue with code in the preceding
commit was that the destination offsets in the send buffer and the source offsets
in the receive buffer didn't advance between successive halo layers, so that,
e.g., halo layer 2 overwrote halo layer 1 in the send buffer, and halo layer 2
was unpacked both to halo layer 1 and halo layer 2.

Now, when creating the compactHaloLayer arrays, the indices in the send and
receive buffers are advanced by the size of the preceding halo layer for each
successive halo layer.