Separate srcList and dstList elements from compactHaloInfo into their own arrays
The indices of elements to move between field arrays and MPI buffers were
previously built into the compactHaloInfo array. The downside to this arrangement
is that the send information had to be skipped over with loops in order to get
to the receive information.

Now, the send and receive element lists have been split into their own arrays.