Update halo code to 21fd87be from atmosphere/acc_group_halos branch
This commit updates the mpas_halo.F and mpas_halo_types.inc with modifications
that are intended to make it easier to implement GPU-direct communication in
the mpas_halo_exch_group_full_halo_exch routine.

An attempt to merge the atmosphere/acc_group_halos branch resulted in a large
number of conflicts due to difference in the dynamics, and since the goal with
this commit is to simply update the halo code, it was deemed easier to simply
commit updated versions of files.