Enable perturbation of initial temperature field using methodology from CAM
This commit implements the capability to add a pseudo-random perturbation to
the initial theta_m field following what is done in the Community Atmosphere
Model (CAM).

Whether or not initial perturbations are added depends on the value of a new
namelist option, config_pertlim, which as been added to a new namelist group
&initial_perturbation. If config_pertlim is 0.0, no initial perturbations
are added; otherwise, the theta_m field is modified on a pointwise basis
through multiplication with a factor (1 + x), where x is a pseudo-random
number in the interval (-pertlim, perlim].

By default, config_pertlim = 0.0, and no initial perturbations are added.

For the same value of config_pertlim, results should be bitwise identical
across executions of the model on different MPI task counts, since the pseudo-
random number generator is seeded independently for each column in a way
that does not depend on, e.g., time of day, horizontal domain decomposition,
etc.