Implement 'in_defaults' attribute for streams in Registry.xml
The 'in_defaults' attribute for a stream functions -- if present -- may be
either "true" or "false", indicating, respectively, whether the stream should
appear in the default streams.<CORE> file.

Omitting the 'in_defaults' attribute is the same as setting in_defaults="true"
for the stream.

Note: In the modifications to streams_gen.c, the condition

    ( write_member == 1 || (write_member == -1 && write_stream) )

has been changed to

    ( write_member != 0 && write_stream )

since we only want to write the stream if write_stream == 1 and write_member
can only take on values 1 (the member is writeable), 0 (the member is not
writeable), or -1 (the member didn't contain any specified attributes).