Ensure that 'scale' scratch array exists on device before monotonic transport
The 'scale' scratch array is allocated on the host before each call to
atm_advance_scalars_mono_work, and deallocated on the host after each call.
To ensure that the array is also present on device -- and since the contents
of the array do not need to be initialized -- this commit adds
'acc enter data create' and `acc exit data delete` directives around the call
to atm_advance_scalars_mono_work.