Implement support for reading and writing through the SMIOL library
This commit updates derived-type definitions to include SMIOL-specific members,
and it adds calls to the SMIOL Fortran API in the mpas_io module to read and
write dimensions, variables, and attributes through the SMIOL library.

The new code added in this commit is only active when the MPAS_SMIOL_SUPPORT
macro is defined at build time.