Split all time-invariant fields into a separate "invariant" input stream
In the atmosphere core's Registry.xml file, all time-invariant input fields have
been placed in a new "invariant" immutable input stream, and the "input" and
"restart" streams include the "invariant" stream.

Since the "invariant" stream is never used as the mesh input stream, the default
input file ("invariant.nc") for the stream does not need to be present in the
run directory, and the "input" and "restart" streams will expect to have all of
the invariant fields included in their files.