Check for input_interval of invariant stream in setup of packages and mesh stream
The atm_setup_packages and atm_get_mesh_stream routines now check the
streams.atmosphere file to see whether it contains an "invariant" stream, and
whether that invariant stream has an input_interval value of "initial_only"; if
so, the "no_invariant_stream" package is deactivated and the "invariant" stream is
used as the mesh stream. When the no_invariant_stream package is deactivated,
neither the "input" stream nor the "restart" stream will read time-invariant
fields from their files, with the expectation that these fields will have been
read from the file associated with the "invariant" stream.