Permit variable number of levels for upper-level horizontal Laplacian filter
The horizontal Laplacian that is used in the upper-level damping layer was
hardwired to be applied over the topmost 3 layers.  This update allows for
the namelist specification of the number of topmost layers over which it is
applied along with a linear ramping of the viscosity used in the Laplacian.

NB: This commit contains an intentional off-by-one indexing bug.