Add logic to use 'surface' stream reference_time when setting 'sfcbdy' alarm
In order for the 'sfcbdy' alarm to ring at the same times as the alarm that
triggers the reading of the 'surface' stream, the mpas_atmphys_manager module
now checks the reference_time attribute for the 'surface' stream when setting up
the 'sfcbdy' alarm.

In case of a restart simulation, however, the 'sfcbdy' alarm is immediately
cleared after having been set up, since we expect the state read from the
restart file to contain exactly what should be used for the first model time
step.