WIP: Correct issue in writing new attributes to existing output files with PIO
When writing an output stream to an existing file, stream writes will fail with
the error

 < ... error message ... >

when using the PIO library.

It appears that opening existing files for writing places those files in netCDF
data mode, and it is therefore necessary to switch to define mode before writing
new attributes.

This commit adds logic for writing scalar integer attributes to switch into
define mode, write the attribute, and to finally switch back to data mode if
writing to an existing output file.