Provide consistency in setting of prevRingTime for alarms when changing ref time
This commit aims to provide consistency in the ringing behavior of alarms after
a call to mpas_adjust_alarm_to_reference_time to alter the reference time for a
recurring alarm.

Previously, when an alarm's reference was set to a time that differed from the
current time by some multiple of the alarm's interval, the prevRingTime for the
alarm became the current time. As such, a query of the alarm's status would show
that it was ringing.

However, in cases where the reference was not different from the current time by
an integer multiple of the alarm interval, the prevRingTime was set to the time
that was earlier than the current time by less than one alarm interval. The
result was that a query of the alarm's status would show that the alarm was not
yet ringing.

Now, the prevRingTime for an alarm is always at least one full alarm interval
before the current time. Whether the difference between the current time and the
reference time is evenly divisible by the alarm's ring interval or not, a query
of the alarm's status will always show that it is ringing.

Note that the changes in this commit only affect alarms when the clock is
running forward in time.